Fix remaining clippy warnings and exclude benchmarks from CI
- Remove unused PoolExt imports from service test modules
- Fix collapsible_match patterns in streaming tests
- Fix field_reassign_with_default in config tests
- Fix if_same_then_else in bash tool tests
- Add clippy allow directive for database benchmarks
- Update CI to exclude benchmarks from clippy (sqlx macros need DATABASE_URL)

All 172 tests passing.

+++ src/db/repository/message.rs
@@ -232,7 +232,7 @@ mod tests {
                 session.id,
                 "user".to_string(),
                 format!("Message {}", i),
-                i as i32 + 1,
+                i + 1,
             );
             message_repo
                 .create(&msg)
